requirements.txt
numpy
Pillow
gunicorn

